improve coverage DiGCL related
