+++ torch_geometric_signed_directed/utils/directed/__init__.py
@@ -1,6 +1,6 @@
 from .prob_imbalance_loss import Prob_Imbalance_Loss
 from .get_magnetic_Laplacian import get_magnetic_Laplacian
-from .get_adjs_DiGCN import get_appr_directed_adj, get_second_directed_adj, cal_fast_appr
+from .get_adjs_DiGCN import get_appr_directed_adj, get_second_directed_adj, cal_fast_appr, fast_appr_power
 from .DiGCL_utils import drop_feature, pred_digcl
 from .features_in_out import directed_features_in_out
 from .meta_graph_generation import meta_graph_generation
